fix: Safely iterate through mutating collection
When deleting multiple photos, iterate through
the array of indexes using a while loop instead
of a for loop.  Though not consistent, the call
to collectionView.indexPathsForSelectedItems()
does not always return in descending row order.
This means that the penultimate index can be
removed from the pin's pictures array before the
last index, creating an index out of range error.
The while loop repeatedly calls for a new array
of index paths for each iteration.

Fixes crash reported by Udacity reviewer.